Refactor skills data move to data folder
