+++ data/skills.ts
@@ -1,0 +1,78 @@
+import { Skill } from "types/Sections";
+
+const skills: Skill[] = [
+  {
+    id: 1,
+    icon: "/images/skills/web.png",
+    name: "Web Design & Development",
+    technologies: ["HTML", "CSS", "JavaScript", "TypeScript"],
+  },
+  {
+    id: 2,
+    icon: "/images/skills/angular.png",
+    name: "Frontend Javascript Frameworks",
+    technologies: ["Angular", "React", "Vue.js", "Svelte"],
+  },
+  {
+    id: 3,
+    icon: "/images/skills/nodejs.png",
+    name: "Backend Javascript Frameworks",
+    technologies: ["Node.js", "Express.js", "NestJS"],
+  },
+  {
+    id: 4,
+    icon: "/images/skills/flutter.png",
+    name: "Cross Platform App Development",
+    technologies: ["Dart", "Flutter", "Firebase"],
+  },
+  {
+    id: 5,
+    icon: "/images/skills/python.png",
+    name: "Scripting Frameworks",
+    technologies: ["Python", "Rust", "Selenium", "Puppeteer"],
+  },
+  {
+    id: 6,
+    icon: "/images/skills/database.png",
+    name: "Database Management System",
+    technologies: ["Postgres", "TypeORM", "MongoDB", "Mongoose"],
+  },
+  {
+    id: 7,
+    icon: "/images/skills/android.png",
+    name: "Android App Development",
+    technologies: ["Android Studio", "Java", "Kotlin", "XML"],
+  },
+  {
+    id: 8,
+    icon: "/images/skills/sketch.png",
+    name: "UI/UX Design",
+    technologies: ["Adobe XD", "Figma", "Sketch"],
+  },
+  {
+    id: 9,
+    icon: "/images/skills/tensorflow.png",
+    name: "Data Processing & Machine Learning",
+    technologies: ["R", "scikit-learn", "TensorFlow", "Keras"],
+  },
+  {
+    id: 10,
+    icon: "/images/skills/seo.png",
+    name: "Search Engine Optimization",
+    technologies: ["PageRank", "Backlinks", "Google Analytics"],
+  },
+  {
+    id: 11,
+    icon: "/images/skills/photography.png",
+    name: "Photography & Cinematography",
+    technologies: ["Adobe Photoshop", "Adobe Lightroom"],
+  },
+  {
+    id: 12,
+    icon: "/images/skills/fcpx.png",
+    name: "Photo & Video Post-Processing",
+    technologies: ["Adobe After Effects", "Final Cut Pro X"],
+  },
+];
+
+export default skills;
